+++ packages/react/src/experimental/Forms/EntitySelect/types.ts
@@ -24,7 +24,7 @@ export type EntitySelectNamedGroup = {
   groupType?: "avatar" | "team"
 }
 
-interface EntitySelectCommonProps
+interface EntitySelectCommonProps<T>
   extends Omit<PopoverProps, "children" | "modal">,
     Pick<
       InputFieldProps<string>,
@@ -36,7 +36,6 @@ interface EntitySelectCommonProps
       | "hint"
       | "hideLabel"
       | "maxLength"
-      | "value"
       | "disabled"
       | "placeholder"
       | "loading"
@@ -68,6 +67,7 @@ interface EntitySelectCommonProps
   onCreate?: (partialName: string) => void
   onCreateLabel?: string
   actions?: Action[]
+  value?: T
 }
 
 export type FlattenedItem = {
@@ -77,18 +77,20 @@ export type FlattenedItem = {
     subItems?: EntitySelectSubEntity[]
   }
 }
-export interface EntitySelectSingleProps extends EntitySelectCommonProps {
+
+export interface EntitySelectSingleProps<T> extends EntitySelectCommonProps<T> {
   onSelect: (entity: EntitySelectEntity | null) => void
   singleSelector: true
 }
 
-export interface EntitySelectMultipleProps extends EntitySelectCommonProps {
+export interface EntitySelectMultipleProps<T>
+  extends EntitySelectCommonProps<T> {
   onSelect: (entities: EntitySelectEntity[]) => void
   singleSelector: false | undefined
 }
 
-export type EntitySelectProps =
-  | EntitySelectSingleProps
-  | EntitySelectMultipleProps
+export type EntitySelectProps<T> =
+  | EntitySelectSingleProps<T>
+  | EntitySelectMultipleProps<T>
 
 export type EntityId = number | string